prettier và sửa model

+++ server/src/controllers/OrganizationController.ts
@@ -1,15 +1,10 @@
 import { Request, Response } from 'express';
 import * as jwt from 'jsonwebtoken';
 import * as dotenv from 'dotenv';
-import {
-  GeneralResponse,
-  commonResponse,
-} from '../utilities/CommonResponse';
-import {
-  Organization,
-  OrganizationAttributes,
-} from '../models/organization';
+import { GeneralResponse, commonResponse } from '../utilities/CommonResponse';
+import { Organization, OrganizationAttributes } from '../models/organization';
 import { organizationMapper } from '../mapper/OrganizationMapper';
+import { randomId } from '../utilities/random';
 dotenv.config();
 const secretKey = process.env.SECRETKEY as string;
 
@@ -21,9 +16,7 @@ export const listOrganization = async (
     const organizationsCurrent = await Organization.findAll({
       where: { status: 0 },
     });
-    const organizations = await organizationMapper(
-      organizationsCurrent,
-    );
+    const organizations = await organizationMapper(organizationsCurrent);
 
     if (organizations.length > 0) {
       const response: GeneralResponse<{
@@ -31,8 +24,7 @@ export const listOrganization = async (
       }> = {
         status: 200,
         data: {
-          organizations:
-            organizations as unknown as OrganizationAttributes[],
+          organizations: organizations as unknown as OrganizationAttributes[],
         },
         message: 'Lấy danh sách tổ chức thành công',
       };
@@ -99,32 +91,39 @@ export const createOrganization = async (
 ): Promise<void> => {
   try {
     const token = req.headers.authorization?.split(' ')[1];
+
     if (!token) {
       res.status(401).json({ message: 'Unauthorized' });
       return;
     }
-
-    const decodedToken = jwt.verify(
-      token,
-      secretKey,
-    ) as jwt.JwtPayload;
+    const decodedToken = jwt.verify(token, secretKey) as jwt.JwtPayload;
     const userId = decodedToken.id;
+    const userRole = decodedToken.role_id;
+
+    if(userRole !== 1 ){
+      res.status(401).json({ message: 'Unauthorized' });
+      return;
+    }
     if (userId) {
+      let organizationId = randomId();
       const body = {
+        orgId: organizationId,
         name: req.body.name as string,
         description: req.body.description as string,
         location: req.body.location as string,
         creator: userId as number,
-        status: 1,
+        status: 2,
         created_at: new Date(),
         updated_at: new Date(),
       };
+      console.log(body);
+      
       await Organization.destroy({ where: { creator: userId } });
       const result = await Organization.create(body);
       if (result) {
         const response: GeneralResponse<{}> = {
           status: 200,
-          data: null,
+          data: body,
           message: 'Tạo tổ chức thành công',
         };
         commonResponse(req, res, response);
